textured works, but the colors?

+++ include/raycast.h
@@ -6,7 +6,7 @@
 /*   By: deordone <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/19 10:54:55 by deordone          #+#    #+#             */
-/*   Updated: 2024/09/20 10:53:26 by santito          ###   ########.fr       */
+/*   Updated: 2024/09/22 18:51:43 by santito          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,24 +23,24 @@ typedef struct s_ray
 	t_ivec			pos;
 	t_ivec			step;
 
-	float			wall_dist;
-	float			wall_start;
-	float			wall_end;
-	float			wall_height;
+	double			wall_dist;
+	double			wall_start;
+	double			wall_end;
+	double			wall_height;
 	int32_t			side;
 }					t_ray;
 
 typedef struct s_wall
 {
-	float			dist;
+	double			dist;
 	int				height;
 	int				start;
 	int				end;
 	int				color;
-	float			hit;
-	float			step;
+	double			hit;
+	double			step;
 	t_ivec			tex;
-	float			tex_pos;
+	double			tex_pos;
 	mlx_texture_t	*texture;
 }					t_wall;
